refactor(mdx): update mdx components
